Move mapDispatchToProps from disco/Addon to core/installAddon
